Replace DatabaseURL with DatabaseConfig to support unsafe characters
It's possible that `password` can contain any characters that make the
whole URL unsafe without being quoted. Instead of asking all clients to
quote their URL components, this commit replaces the whole usage of
DatabaseURL with DatabaseConfig which is a `dataclass` that contains all
the necessary components to establish a database connection.

To preserve some extent of backwards compatibility, the new
`DatabaseConfig` class has `to_url` and `from_url` methods that mimick
the previous `DatabaseURL` behaviour, should that be needed.